datagen fixes

+++ src/helper.cpp
@@ -66,14 +66,14 @@ Square msb(Bitboard b) {
 
 #endif
 
-int popcount(Bitboard mask) {
+unsigned popcount(Bitboard mask) {
 #if defined(_MSC_VER) || defined(__INTEL_COMPILER)
 
-    return (uint8_t)_mm_popcnt_u64(mask);
+    return (unsigned)_mm_popcnt_u64(mask);
 
 #else  // Assumed gcc or compatible compiler
 
-    return __builtin_popcountll(mask);
+    return (unsigned)__builtin_popcountll(mask);
 
 #endif
 }
